files-view: Set NAUTILUS_SCRIPT_SELECTED_FILE_PATHS in starred:/
We have been calling is_local_or_fuse() on the current directory to
know if we can pass local paths for scripts in as an environment
variable.

This is wrong because the current directory may not be the actual
location of the files. That's why a workaround has been introduced for
search directories in commit 8fdd3792ed4e1fb95e6c4f2adc4c8631c27bb2d3
But the same problem still happens in the starred:/ directory.

In the case of paths passed as command line parameters, we call
is_local_or_fuse() directly on the files, so there is no need for any
workaround for search and starred. This enhancement had been proposed
in https://bugzilla.gnome.org/show_bug.cgi?id=341657, and introduced
in commit 34d594099868445b1c2d4f16da3bcdf05f914100.

However, this enhancement was not applied to the envar case. Let's do
something equivalent now.