preferences: Drop executable-text-activation setting
Whe the ASK option gone, the only alternative to the default is RUN.

This option is not safe: it's too easy to accidentally run programs.

Now we have a "Run as a Program" action in the context menu, which is
available by default (no configuration required) and safe (clearly
labeled, intentionally chosen).

So, remove the RUN option is no longer necessary, and with it we can
remove the preference from both the UI and GSettings schemas.

Closes https://gitlab.gnome.org/GNOME/nautilus/-/issues/1700
Discussed in https://gitlab.gnome.org/GNOME/nautilus/-/issues/443